requirements.txt
asgiref==3.11.0
Django==5.2.8
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gprof2dot==2025.4.14
jalali_core==1.0.0
jdatetime==5.2.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
